refactor: share DateHelper in common package
