+++ apps/backend/Backend/directusExtensions/directus-extension-rocket-meals-bundle/src/food-feedback-report-schedule/__tests__/TestFoodFeedbackReportScheduleGetReferenceDate.ts
@@ -1,7 +1,7 @@
 // small jest test
 import {describe, it} from '@jest/globals';
 import {DatabaseTypes} from "repo-depkit-common";
-import {DateHelper} from "../../helpers/DateHelper";
+import {DateHelper} from "repo-depkit-common";
 import {ReportSchedule} from "../ReportSchedule";
 import {FoodofferDateType} from "../../food-sync-hook/FoodParserInterface";
