+++ apps/backend/Backend/directusExtensions/directus-extension-rocket-meals-bundle/src/helpers/EnvVariableHelper.ts
@@ -1,4 +1,4 @@
-import {DateHelperTimezone} from "./DateHelper";
+import {DateHelperTimezone} from "repo-depkit-common";
 
 export enum SyncForCustomerEnum {
     TEST = "Test",
