+++ packages/common/package.json
@@ -6,5 +6,8 @@
   "private": true,
   "devDependencies": {
     "typescript": "*"
+  },
+  "dependencies": {
+    "moment-timezone": "^0.5.48"
   }
 }
